pyproject.toml
[project]
name = "bmadnotion"
version = "0.2.0"
description = "Sync BMAD project artifacts to Notion - Pages and Databases"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [
    { name = "Li Bo", email = "markdown@outlook.com" }
]
keywords = [
    "bmad",
    "notion",
    "sync",
    "markdown",
    "project-management",
    "agile",
    "sprint-tracking"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Documentation",
    "Topic :: Office/Business :: Groupware",
    "Typing :: Typed",
]
dependencies = [
    "marknotion>=0.2.0",
    "notion-client>=2.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
bmad = "bmadnotion.cli:cli"
bmadnotion = "bmadnotion.cli:cli"  # Alias for backwards compatibility

[project.urls]
Homepage = "https://github.com/li3p/bmadnotion"
Repository = "https://github.com/li3p/bmadnotion"
Documentation = "https://github.com/li3p/bmadnotion#readme"
Issues = "https://github.com/li3p/bmadnotion/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bmadnotion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
markers = [
    "integration: marks tests as integration tests (require Notion API)",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.coverage.run]
source = ["src/bmadnotion"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.uv.sources]
marknotion = { path = "../marknotion", editable = true }
